STAC-14483 Centralize code
